chore(sonar): remove NTLM dead store + two redundant casts
- ntlm.cpp: drop the final 'k += 4' (S1854) - k is a write index only; the HMAC uses temp_len
  (which already includes the trailing Z(4)), so k is never read after. Verified vs MS-NLMP layout;
  native_smb (all 6 SMB suites incl. NTLM auth) green.
- forward.cpp / mtconnect.cpp: drop redundant (uint32_t) casts (S1905) - every operand is already
  uint32_t, so no narrowing.

+++ src/services/forward/forward.cpp
@@ -147,7 +147,7 @@ bool rate_gate(uint32_t &window_start, uint16_t &count, uint16_t rate_cap)
     if (rate_cap == 0)
         return false; // unlimited
     uint32_t now = fwd_now();
-    if ((uint32_t)(now - window_start) >= 1000)
+    if ((now - window_start) >= 1000)
     {
         window_start = now;
         count = 0;
